move make_trace & make_instructions to the walker
